refactor(#118): 글 상세 디자인 완성 및 코드 refactor
